feat: implement FilterService, add filtering tests
Took 32 minutes
