+++ api/src/main/java/demo/controller/FilterController.java
@@ -5,7 +5,7 @@ import demo.controller.dto.UserFilterDto;
 import demo.mapper.UserFilterMapper;
 import demo.model.User;
 import demo.model.UserFilter;
-import demo.service.FilterService;
+import demo.service.validation.filter_validator.FilterService;
 import demo.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
